+++ tests/test_eventable.py
@@ -19,6 +19,9 @@ from wyoming.info import (
     HandleProgram,
     IntentModel,
     IntentProgram,
+    Identity,
+    IdentityModel,
+    IdentityProgram,
     MicProgram,
     Satellite,
     SndProgram,
@@ -115,6 +118,25 @@ TEST_DATA: Dict[str, Dict[str, Any]] = {
                     )
                 ],
                 supports_synthesize_streaming=True,
+            )
+        ],
+        "identity": [
+            IdentityProgram(
+                name=TEST_NAME,
+                attribution=TEST_ATTRIBUTION,
+                installed=True,
+                description=TEST_DESCRIPTION,
+                version=TEST_VERSION,
+                models=[
+                    IdentityModel(
+                        name=TEST_NAME,
+                        attribution=TEST_ATTRIBUTION,
+                        installed=True,
+                        description=TEST_DESCRIPTION,
+                        version=TEST_VERSION,
+                        identities=[Identity(TEST_SPEAKER)],
+                    )
+                ],
             )
         ],
         "handle": [
@@ -268,6 +290,18 @@ TEST_DATA: Dict[str, Dict[str, Any]] = {
     "SynthesizeStop": {},
     "SynthesizeStopped": {},
     "Synthesize": {"text": TEST_TEXT, "voice": TEST_VOICE, "context": TEST_CONTEXT},
+    # identity
+    "Identify": {
+        "name": TEST_NAME,
+        "names": [TEST_SPEAKER],
+        "context": TEST_CONTEXT,
+    },
+    "Identified": {
+        "name": TEST_SPEAKER,
+        "confidence": 0.9,
+        "context": TEST_CONTEXT,
+    },
+    "NotIdentified": {"context": TEST_CONTEXT},
     # timers
     "TimerStarted": {"id": TEST_ID, "total_seconds": 100},
     "TimerUpdated": {"id": TEST_ID, "total_seconds": 100, "is_active": True},